job-hunter-agent/pyproject.toml
[project]
name = "job-hunter-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.5.0",
    "firecrawl-py>=4.6.0",
    "python-dotenv>=1.2.1",
]
